- Added define __bootstrap to replace section usage with string constant - Replaced SECTION( ".text.boot" ) by new define __bootstrap - Call correct function within boot_virt_map depending on supported mode - Fixed union type definitions for lpae - Splitted up lpae table context into two different types - Fixed defines extracting the index from a virtual address from lpae - Changed physical address to 64bit integer with conversion to 32bit if necessary - Added working version of initial mmu setup with lpae - Removed unnecessary duplicated function declaration

+++ include/arch/arm/v7/boot/mm/virt/long.h
@@ -24,6 +24,6 @@
 #include <kernel/boot/mm/virt.h>
 
 void boot_virt_setup_long( uintptr_t );
-void boot_virt_map_long( uintptr_t, uintptr_t );
+void boot_virt_map_long( uint64_t, uintptr_t );
 
 #endif
